refactor(analysis): extract content cognitive scoring
